tdd: support with VS202x
Enable TDD compatibility for s32 driver modules when building with VS202x.

+++ drivers/can/can_mcux_flexcan.c
@@ -318,7 +318,7 @@ static int mcux_flexcan_start(const struct device *dev)
 		timing.frJumpwidth = data->timing_data.sjw - 1U;
 		timing.fphaseSeg1 = data->timing_data.phase_seg1 - 1U;
 		timing.fphaseSeg2 = data->timing_data.phase_seg2 - 1U;
-		timing.fpropSeg = data->timing_data.prop_seg;
+		timing.fpropSeg = (uint8_t)data->timing_data.prop_seg;
 		FLEXCAN_SetFDTimingConfig(base, &timing);
 
 		FLEXCAN_EnterFreezeMode(base);
@@ -1265,7 +1265,7 @@ static int mcux_flexcan_init(const struct device *dev)
 #ifdef CONFIG_CAN_MCUX_FLEXCAN_FD
 	if (config->flexcan_fd) {
 		flexcan_config.timingConfig.frJumpwidth = data->timing_data.sjw - 1U;
-		flexcan_config.timingConfig.fpropSeg = data->timing_data.prop_seg;
+		flexcan_config.timingConfig.fpropSeg = (uint8_t)data->timing_data.prop_seg;
 		flexcan_config.timingConfig.fphaseSeg1 = data->timing_data.phase_seg1 - 1U;
 		flexcan_config.timingConfig.fphaseSeg2 = data->timing_data.phase_seg2 - 1U;
